Fix python wrapper
