+++ python/all_mrpt_math1.cpp
@@ -1,11 +1,12 @@
 #include "src/mrpt/math/CAtan2LookUpTable.cpp"
 #include "src/mrpt/math/CMatrixB.cpp"
+#include "src/mrpt/math/CMatrixDynamic_2.cpp"
 #include "src/mrpt/math/CMatrixDynamic_1.cpp"
 #include "src/mrpt/math/CMatrixDynamic.cpp"
 #include "src/mrpt/math/CMatrixF.cpp"
 #include "src/mrpt/math/CMatrixFixed_1.cpp"
 #include "src/mrpt/math/CMatrixFixed_2.cpp"
 #include "src/mrpt/math/CMatrixFixed_3.cpp"
-#include "src/mrpt/math/CMatrixFixed_4.cpp"
 #include "src/mrpt/math/CMatrixFixed.cpp"
+#include "src/mrpt/math/MatrixVectorBase.cpp"
 #include "src/mrpt/math/CPolygon.cpp"
