+++ python/python.conf
@@ -213,6 +213,7 @@
 -class mrpt::math::MatrixBase<double,mrpt::math::CMatrixFixed<double, 7, 7>>
 -class mrpt::math::MatrixBase<unsigned char,mrpt::math::CMatrixDynamic<unsigned char>>
 -class mrpt::math::MatrixBase<unsigned short,mrpt::math::CMatrixDynamic<unsigned short>>
+-class mrpt::math::MatrixBase<mrpt::math::TPoint3D_<float>,mrpt::math::CMatrixDynamic<mrpt::math::TPoint3D_<float> >>
 -class mrpt::math::MatrixBase<float,mrpt::math::CMatrixDynamic<float>>
 -class mrpt::math::MatrixBase<float,mrpt::math::CMatrixFixed<float, 3, 1>>
 -class mrpt::math::MatrixBase<float,mrpt::math::CMatrixFixed<float, 3, 3>>
@@ -232,6 +233,7 @@
 -class mrpt::math::MatrixVectorBase<float,mrpt::math::CMatrixFixed<float, 3, 1>>
 -class mrpt::math::MatrixVectorBase<float,mrpt::math::CMatrixFixed<float, 4, 1>>
 -class mrpt::math::MatrixVectorBase<unsigned char,mrpt::math::CMatrixDynamic<unsigned char>>
+-class mrpt::math::MatrixVectorBase<mrpt::math::TPoint3D_<float>,mrpt::math::CMatrixDynamic<mrpt::math::TPoint3D_<float> >>
 -class mrpt::math::MatrixVectorBase<double,mrpt::math::CMatrixFixed<double, 2, 2>>
 -class mrpt::math::MatrixVectorBase<double,mrpt::math::CMatrixFixed<double, 7, 7>>
 -class mrpt::math::MatrixVectorBase<double,mrpt::math::CMatrixDynamic<double>>
